new-ai-data-science-team/requirements.txt
# Core Dependencies
python-dotenv>=0.19.0
pydantic>=2.0.0

# Data Collection
scrapy>=2.5.0
beautifulsoup4>=4.9.3
selenium>=4.0.0
great-expectations>=0.15.0

# Data Storage
pymongo>=4.0.0
psycopg2-binary>=2.9.0
redis>=4.0.0

# Data Processing
pandas>=1.4.0
numpy>=1.21.0
dask>=2022.1.0
vaex>=4.0.0
pandas-profiling>=3.2.0
dedupe>=2.0.0

# Feature Engineering
scikit-learn>=1.0.0
feature-engine>=1.2.0
category-encoders>=2.3.0

# Analysis
statsmodels>=0.13.0
scipy>=1.7.0
plotly>=5.5.0
streamlit>=1.8.0

# ML Libraries
xgboost>=1.5.0
lightgbm>=3.3.0
tensorflow>=2.8.0
torch>=1.10.0
optuna>=2.10.0

# Model Interpretability
shap>=0.40.0
lime>=0.2.0
alibi>=0.7.0
mlflow>=1.24.0

# Monitoring & Logging
prometheus-client>=0.13.0
elasticsearch>=7.17.0
sentry-sdk>=1.5.0

# Development Tools
pytest>=7.0.0
hypothesis>=6.36.0
coverage>=6.3.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.931
pre-commit>=2.17.0
sphinx>=4.4.0
mkdocs>=1.2.0

# Container
docker>=5.0.0
docker-compose>=1.29.0
